+++ tests/processtests.cpp
@@ -31,6 +31,12 @@ TEST_F(ProcessTest, Start)
     ASSERT_FALSE(m_proc->hasCompleted());
 }
 
+TEST_F(ProcessTest, Hardware)
+{
+    ASSERT_TRUE(m_proc->getRAMUsage() > 0);
+    ASSERT_NO_THROW(m_proc->getCPUUsage());
+}
+
 TEST_F(ProcessTest, Kill)
 {
     ASSERT_TRUE(m_proc->kill());